Override default value formatter, courtesy of @snakefoot
